totalSpaceRequired changed to match typical use
The user might not have a list of name/size pairs,
but they will always have a list of name/
(alignment/size) pairs in order to allocate space
for the variables.  Thus, totalSpaceRequired was
changed to take a list of the latter rather than
of the former to make it easier to use.