Added better support for the sizeof operator
Now Privateer attempts to see if it is being applied to a global variable or
a local static variable.  If so, it replaces the expression with the known
size of the variable.  This is needed because in some cases the original
variable is replaced with a type whose size is not known at compile-time,
such as an array with unspecified length.